File Make.cp.irix replaced by the common Make.irix machine file In CPV/Makefile FPMD/Makefile LDFLAGS now used like in PW/Makefile
git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@121 c92efa57-630b-4861-b058-cf58834340f0

+++ install/sizeof.c
@@ -8,13 +8,13 @@
 int main()
 {
 #if defined __FFTW
-  printf("SIZEOF ( fftw_plan * ). . = %d\n",sizeof(fftw_plan *));
-  printf("SIZEOF ( FFTW_COMPLEX * ) = %d\n",sizeof(FFTW_COMPLEX *));
-  printf("SIZEOF ( FFTW_COMPLEX ) . = %d\n",sizeof(FFTW_COMPLEX));
+  printf("SIZEOF ( fftw_plan * ). . = %d\n",(int)sizeof(fftw_plan *));
+  printf("SIZEOF ( FFTW_COMPLEX * ) = %d\n",(int)sizeof(FFTW_COMPLEX *));
+  printf("SIZEOF ( FFTW_COMPLEX ) . = %d\n",(int)sizeof(FFTW_COMPLEX));
 #endif
-  printf("SIZEOF ( void * ) . . . . = %d\n",sizeof(void *));
-  printf("SIZEOF ( double ) . . . . = %d\n",sizeof(double));
-  printf("SIZEOF ( float ). . . . . = %d\n",sizeof(float));
-  printf("SIZEOF ( int ). . . . . . = %d\n",sizeof(int));
-  printf("SIZEOF ( char ) . . . . . = %d\n",sizeof(char));
+  printf("SIZEOF ( void * ) . . . . = %d\n",(int)sizeof(void *));
+  printf("SIZEOF ( double ) . . . . = %d\n",(int)sizeof(double));
+  printf("SIZEOF ( float ). . . . . = %d\n",(int)sizeof(float));
+  printf("SIZEOF ( int ). . . . . . = %d\n",(int)sizeof(int));
+  printf("SIZEOF ( char ) . . . . . = %d\n",(int)sizeof(char));
 }
